job-pulse/requirements.txt
# Core dependencies
requests>=2.31.0
feedparser>=6.0.10
python-dateutil>=2.8.2
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
flask>=3.0.0

# Resume parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Google Sheets export (optional)
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

